resolve issue #682
